security: remover application.properties del tracking
- Archivo removido del control de versiones
- Agregado a .gitignore para protección futura
- Mantener archivo local para configuración
