+++ Usuario/backend/backend/src/main/java/com/usuario/backend/security/oauth2/OAuth2AuthenticationFailureHandler.java
@@ -1,23 +1,28 @@
-package com.usuario.backend.security;
-
-import com.usuario.backend.model.Usuario;
-import com.usuario.backend.service.UsuarioService;
+package com.usuario.backend.security.oauth2;
+import com.usuario.backend.model.entity.Usuario;
+import com.usuario.backend.service.user.UsuarioService;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+import org.springframework.web.util.UriComponentsBuilder;
+import com.usuario.backend.security.jwt.JwtTokenProvider;
 
 import java.io.IOException;
+import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.Map;
 
 @Component
-public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
+public class OAuth2AuthenticationFailureHandler extends SimpleUrlAuthenticationSuccessHandler {
 
     private static final Logger logger = LoggerFactory.getLogger(OAuth2AuthenticationSuccessHandler.class);
 
@@ -26,6 +31,9 @@ public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationS
 
     @Autowired
     private UsuarioService usuarioService;
+    
+    @Value("${app.frontend.url:http://localhost:5173}")
+    private String frontendUrl;
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication)
@@ -43,17 +51,13 @@ public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationS
             String email = (String) attributes.get("email");
             String name = (String) attributes.get("given_name");
             String lastName = (String) attributes.get("family_name");
-            
-            // 🎯 OBTENER IMAGEN DE GOOGLE
-            String pictureUrl = (String) attributes.get("picture");
-            logger.info("Google profile picture URL: {}", pictureUrl);
-            
+
             logger.info("User authenticated with OAuth2: {}", email);
 
             // Verificar dominio institucional
             if (!email.endsWith("@tecsup.edu.pe")) {
                 logger.error("Error: Email domain not allowed: {}", email);
-                response.sendRedirect("http://localhost:5173/login?error=domain_not_allowed");
+                response.sendRedirect(frontendUrl + "/login?error=" + URLEncoder.encode("Solo se permiten correos con dominio @tecsup.edu.pe", StandardCharsets.UTF_8));
                 return;
             }
 
@@ -62,19 +66,13 @@ public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationS
                 Usuario existingUser = usuarioService.findByCorreoInstitucional(email);
 
                 if (existingUser == null) {
-                    // 🔧 CREAR NUEVO USUARIO CON IMAGEN DE GOOGLE
+                    // Crear nuevo usuario con los datos básicos
                     Usuario newUser = new Usuario();
                     newUser.setCorreoInstitucional(email);
                     newUser.setNombre(name != null ? name : "");
                     newUser.setApellidos(lastName != null ? lastName : "");
                     newUser.setRol("ESTUDIANTE");
-                    
-                    // 🎯 GUARDAR IMAGEN DE GOOGLE
-                    if (pictureUrl != null && !pictureUrl.isEmpty()) {
-                        logger.info("Guardando imagen de Google para nuevo usuario: {}", pictureUrl);
-                        newUser.setProfileImageUrl(pictureUrl);
-                    }
-                    
+
                     // Generar código de usuario basado en el email
                     String codigo = email.split("@")[0];
                     newUser.setCodigo(codigo);
@@ -86,49 +84,17 @@ public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationS
                     newUser.setDepartamentoId(1L); // Tecnología Digital
 
                     // Registrar usuario
-                    Usuario usuarioCreado = usuarioService.registrarUsuarioOAuth(newUser);
-                    logger.info("Created new user with Google image for: {} - Image URL: {}", 
-                              email, usuarioCreado.getProfileImageUrl());
-                    
+                    usuarioService.registrarUsuarioOAuth(newUser);
+                    logger.info("Created new user for: {}", email);
                 } else {
-                    // 🔧 ACTUALIZAR USUARIO EXISTENTE CON IMAGEN DE GOOGLE
                     logger.info("User already exists: {}", email);
-                    
-                    // Actualizar información básica
-                    existingUser.setNombre(name != null ? name : existingUser.getNombre());
-                    existingUser.setApellidos(lastName != null ? lastName : existingUser.getApellidos());
-                    
-                    // 🎯 ACTUALIZAR IMAGEN SOLO SI NO TIENE O SI QUIERE LA DE GOOGLE
-                    boolean shouldUpdateImage = false;
-                    
-                    if (pictureUrl != null && !pictureUrl.isEmpty()) {
-                        if (existingUser.getProfileImageUrl() == null || existingUser.getProfileImageUrl().isEmpty()) {
-                            // No tiene imagen, usar la de Google
-                            shouldUpdateImage = true;
-                            logger.info("Usuario sin imagen, usando imagen de Google: {}", pictureUrl);
-                        } else if (existingUser.getProfileImageUrl().contains("googleusercontent.com")) {
-                            // Ya tiene imagen de Google, actualizarla por si cambió
-                            shouldUpdateImage = true;
-                            logger.info("Actualizando imagen de Google existente: {}", pictureUrl);
-                        } else {
-                            // Tiene imagen personalizada, mantenerla
-                            logger.info("Usuario tiene imagen personalizada, manteniéndola: {}", existingUser.getProfileImageUrl());
-                        }
-                        
-                        if (shouldUpdateImage) {
-                            existingUser.setProfileImageUrl(pictureUrl);
-                            Usuario usuarioActualizado = usuarioService.actualizarUsuario(existingUser);
-                            logger.info("Updated user with Google image: {} - New URL: {}", 
-                                      email, usuarioActualizado.getProfileImageUrl());
-                        }
-                    }
                 }
             } catch (Exception e) {
                 logger.error("Error processing user: {}", e.getMessage(), e);
-                response.sendRedirect("http://localhost:5173/login?error=user_processing");
+                response.sendRedirect(frontendUrl + "/login?error=" + URLEncoder.encode("Error al procesar usuario: " + e.getMessage(), StandardCharsets.UTF_8));
                 return;
             }
-            
+
             // Generar token JWT
             String token = null;
             try {
@@ -136,12 +102,12 @@ public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationS
                 logger.info("JWT token generated for user: {}", email);
             } catch (Exception e) {
                 logger.error("Error generating token: {}", e.getMessage(), e);
-                response.sendRedirect("http://localhost:5173/login?error=token_generation");
+                response.sendRedirect(frontendUrl + "/login?error=" + URLEncoder.encode("Error al generar token: " + e.getMessage(), StandardCharsets.UTF_8));
                 return;
             }
 
             // URL frontend con el token
-            String targetUrl = "http://localhost:5173/home?token=" + token;
+            String targetUrl = frontendUrl + "/home?token=" + token;
             logger.info("About to redirect to: {}", targetUrl);
 
             // Redirección al frontend
@@ -150,7 +116,7 @@ public class OAuth2AuthenticationSuccessHandler extends SimpleUrlAuthenticationS
 
         } catch (Exception e) {
             logger.error("General error in OAuth authentication success handler", e);
-            response.sendRedirect("http://localhost:5173/login?error=general_error");
+            response.sendRedirect(frontendUrl + "/login?error=" + URLEncoder.encode("Error general: " + e.getMessage(), StandardCharsets.UTF_8));
         }
     }
 }
