--- Usuario/backend/backend/src/main/resources/application.properties
@@ -1,39 +1,0 @@
-spring.application.name=backend
-spring.datasource.url=jdbc:mysql://localhost:3306/teckbook 
-spring.datasource.username=root
-spring.datasource.password=
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
-
-server.port=8080
-
-# Cloudinary configuration
-cloudinary.cloud-name=
-cloudinary.api-key=
-cloudinary.api-secret=
-# Tamaño máximo para subida de archivos
-spring.servlet.multipart.max-file-size=10MB
-spring.servlet.multipart.max-request-size=10MB
-
-# Usa
-spring.security.oauth2.client.registration.google.client-id=
-spring.security.oauth2.client.registration.google.client-secret=
-spring.security.oauth2.client.registration.google.redirect-uri={baseUrl}/oauth2/callback/google
-spring.security.oauth2.client.registration.google.scope=email,profile
-
-
-# Configuraci n JWT
-app.jwt.secret=BQfe57X5tqMOhhWZSP0vNMHQhMuoUFIcKNJY6xH1iEDEVejZ4vMOAUxgY0DSRPzUFPAyI6t0tDgQNJ0nPxs52g
-app.jwt.expiration-ms=86400000
-
-# Logging detallado
-logging.level.org.springframework.security=DEBUG
-logging.level.org.springframework.security.oauth2=DEBUG
-logging.level.com.usuario.backend.security=DEBUG
-
-# Mostrar mensajes de error detallados
-server.error.include-stacktrace=always
-server.error.include-message=always
